+++ ui/src/types.ts
@@ -1,25 +1,3 @@
-export class PerQuote extends Map<string, Array<Asset>>{ };
-
-export type Assets = Array<Asset>;
-
-export type Accounts = Map<string, Assets>;
-
-export interface Account {
-    averageEntryPrice: number,
-    closedPnl: number,
-    closedQuantity: number,
-    currentMarketValue: number,
-    currentPrice: number,
-    dayPnl: number,
-    isRealTime: boolean,
-    isUnderReorg: boolean,
-    openPnl: number,
-    openQuantity: number,
-    symbol: string,
-    symbolId: number,
-    totalCost: number
-}
-
 export interface Asset {
     averageEntryPrice: number,
     closedPnl: number,
@@ -35,3 +13,9 @@ export interface Asset {
     symbolId: number,
     totalCost: number
 }
+
+export type Assets = Array<Asset>;
+
+export type Accounts = Map<string, Assets>;
+
+export class PerQuote extends Map<string, Array<Asset>>{ };